Moved GradleReport to the api package
